refactor(history): add metadata to manage chat history in some agents
- Add get_metadata and set_metadata functions to manage chat history metadata
- Modify history file structure to include metadata record at the beginning
- Update get_history and get_history_list to filter out metadata records

Example:
```
{
    "role": "metadata",
    "timestamp": "2025-01-14T14:52:13",
    "resume_id": "616bf755-08a5-4b42-b0b5-f41efd0b28ca"
  }
```

The metadata context can be different across different agents.
Some agents' memory systems may be incompatible with other agents' memory systems. This can be specified in the metadata and validated within the agent (though it's not strictly necessary - it won't cause errors if you don't, but the memory just won't be set correctly).