Fix: Check if user conf is up-to-date with defaults at launch
## Problem:
The user (mainly developers of this project) would use git instead of the upgrade script to sync the commits.  After untracked the conf.yaml from git, the user config is no longer synced. Without using the upgrade script, the user config will be out-of-date.

This may lead to errors in configuration validation process.

## Solution:
We need to check if the user configuration is up-to-date with the default configuration every time we launch the backend and update the configuration if needed.

We only compare the conf_version between the default and the user configuration